export internal function fill_copy_indices for usage in other transfer classes
